moving components to organize the project
