+++ src/app/app-routing.module.ts
@@ -1,8 +1,8 @@
 import { Component, NgModule } from '@angular/core';
 import { RouterModule, Routes } from '@angular/router';
-import { HomeComponent } from './components/home/home.component';
-import { FilmesComponent } from './components/filmes/filmes.component';
-import { NaveComponent } from './components/nave/nave.component';
+import { HomeComponent } from './components/pages/home/home.component';
+import { FilmesComponent } from './components/pages/filmes/filmes.component';
+import { NaveComponent } from './components/pages/nave/nave.component';
 
 const routes: Routes = [
   {
