+++ src/app/app.component.html
@@ -1,4 +1,4 @@
 <app-header></app-header>
-<main>
+<section>
   <router-outlet></router-outlet>
-</main>
+</section>
